sizeof

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int main(void)
 	long long int h;
 	float j;
 
-	printf("Size of char: %lu byte(s)\n", sizeof(a));
-	printf("Size of char: %lu byte(s)\n", sizeof(b));
-	printf("Size of char: %lu byte(s)\n", sizeof(g));
-	printf("Size of char: %lu byte(s)\n", sizeof(h));
-	printf("Size of char: %lu byte(s)\n", sizeof(j));
+	printf("Size of char: %lu byte(s)\n", (unsigned long)sizeof(a));
+	printf("Size of char: %lu byte(s)\n", (unsigned long)sizeof(b));
+	printf("Size of char: %lu byte(s)\n", (unsigned long)sizeof(g));
+	printf("Size of char: %lu byte(s)\n", (unsigned long)sizeof(h));
+	printf("Size of char: %lu byte(s)\n", (unsigned long)sizeof(j));
 	return (0);
 }
